requirements.txt
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.0
joblib==1.5.2
kiwisolver==1.4.9
llvmlite==0.46.0
matplotlib==3.10.7
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
orca==1.0
packaging==25.0
pandas==2.3.3
pandas-stubs==2.3.3.251201
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.9.0
shap==0.50.0
six==1.17.0
slicer==0.0.8
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
